refactor: Move minimalSupportedStatuses() to its own file
