+++ tests/DocsSamplesForStatuses.test.ts
@@ -12,6 +12,7 @@ import { Group } from '../src/Query/Group';
 import { StatusNameField } from '../src/Query/Filter/StatusNameField';
 import { StatusTypeField } from '../src/Query/Filter/StatusTypeField';
 import type { StatusCollection } from '../src/StatusCollection';
+import { minimalSupportedStatuses } from '../src/Config/Themes/MinimalThemeCollection';
 import { TaskBuilder } from './TestingTools/TaskBuilder';
 
 function verifyMarkdown(markdown: string) {
@@ -131,7 +132,7 @@ describe('Theme', () => {
     });
 
     describe('Minimal', () => {
-        const statuses = StatusSettingsHelpers.minimalSupportedStatuses();
+        const statuses = minimalSupportedStatuses();
         it('Table', () => {
             verifyStatusesAsMarkdownTable(constructStatuses(statuses));
         });
